Introduce object manager classes for ServiceWorkerContainerHost
To prepare for splitting `ServiceWorkerContainerHost` and
to make the class smaller,
this CL introduces two manager classes that owns and manages
`ServiceWorkerRegistrationObjectHost` and
`ServiceWorkerObjectHost` objects.

This CL doesn't change the behavior nor lifetime, because the
new manager objects corresponds 1:1 to and have the same lifetime
as `ServiceWorkerContainerHost`.

Bug: 336154571
Change-Id: I7602c628dfd746d63c4e012b0264e7004259e820
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5470904
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Commit-Queue: Hiroshige Hayashizaki <hiroshige@chromium.org>
Reviewed-by: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292835}