[URL] Make URLUtilTest pass with the non-special URL feature
URLUtilTest is one of the tests which would fail when we
enable the StandardCompliandNonSpecialSchemeURLParser feature
[^1][^2]:

The test currently uses a non-special URL, like "bogus://google.com".

When the feature is enabled, the host part in non-special URLs can be
correctly parsed. That's why the test would fail with the feature
enabled.

See the following URL parser tester for details:

https://timothygu.me/urltester/#input=bogus%3A%2F%2Fgoogle.com3A%2F%2Fnewtab
hostname:
> whatwg-url : "google.com"
> Chrome: (empty string)

This CL makes the test pass, regardless of whether the feature is
enabled or not, by making the test a parameterized test and adding
expected outputs when the feature is enabled, in preparation for
enabling the feature.

Though the main purpose of this CL is to make the test pass when the
feature is enabled, I'd like to notify a code owner that we are
changing the behavior of non-special URLs via this CL's review. Since
I'm not familiar how `MatchFilters` should work for non-special URLs,
it's recommended for a code owner or a domain expert to review and
potentially do an appropriate action in a follow-up CL, if necessary.

Depending on the severity of a potential breakage, I'd appreciate if
you could file a bug and marking it as blocking for bug 40063064 so we
don't ship the feature until all critical issues are addressed.

[^1]: https://ci.chromium.org/ui/p/chromium/builders/try/linux-rel/1752691/test-results
[^2]: https://docs.google.com/spreadsheets/d/1Pqw1iKXK_lxHj-kLIAeRFs-khFz-BPZDio1W7SgEVE4/

Bug: 40063064
Change-Id: Ia0f2b228f1f34a09b27a0c116ed1724c532977ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5475006
Commit-Queue: Hayato Ito <hayato@chromium.org>
Reviewed-by: Dominic Battre <battre@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292858}