bluetooth: Set up uuids during BluetothClassicDeviceMac construction
Currently, BluetoothClassDeviceMac::GetUUIDs retrieves the latest UUIDs
from the system when it is called, this might not align with how
BluetothAdapterMac uses it as it relies on the snapshot of UUIDs when
the BluetoothClassDeviceMac object is created for comparison of old and
new device objects. This CL fixes it by retrieving and storing the
device UUIDs when the device is constructed, so that the future device
change can be properly detected by comparing UUIDs of the old and new
device objects.

Bug: 326848790
Change-Id: Idb24db57f40ad3cf531ba428de4032cf9ceadf43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5486476
Commit-Queue: Jack Hsieh <chengweih@chromium.org>
Reviewed-by: Matt Reynolds <mattreynolds@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292865}