Move PDM::AddChangeCallback() to ADM
AddChangeCallback() is a way to add an observer for only the next
OnPersonalDataChanged() event as a base::OnceCallback(). This is useful
in cases where the observing class (e.g. the context menu manager) might
be gone by the time the notification is triggered, but something (like
opening the popup) still needs to happen after a change has propagated.

In any case, this is only used by address code, so it is moved to the
ADM and made part of the ADM observer rather than the PDM observer.
The code is moved without changes, though the semantics obviously change
as the AddChangeCallback() only cares about address changes now. The
single user of this mechanism is updated too.

This allows moving the last PDM test to the ADM tests. Consequently, the
PDM unittest file is deleted.

Bug: b/322170538
Change-Id: Ic74af620b837936df83ccc6a75619cd060a307d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5488097
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Cr-Commit-Position: refs/heads/main@{#1292899}