Introduce kGoogleServicesLastSignedInUsername pref
Similar to kGoogleServicesLastSyncingUsername, but written for all
signed-in users, no matter whether they are syncing or not.

Use case: on iOS, if the signed-in account is removed from the device,
a UI is shown to offer the user to sign-in again. This flow pre-fills
the email of the old account in the Gaia sign-in page [1]. Also, the UI
is shown on every startup until the user explicitly accepts or dismisses
it, meaning the email string needs to stay *persisted* until then.
This flow used to be implemented with kGoogleServicesLastSyncingUsername
but syncing users don't exist on iOS anymore, so we need something else.
We could introduce a dedicated pref, but at that point it seems better
to just do it generically.

This CL doesn't fix the iOS UI yet, it just introduces the new pref.

[1] https://crbug.com/40071162#comment31 shows a demo video of the flow
    for a syncing user.

Bug: 40071162
Change-Id: Ie5dead29321da9bf830a2c21ac8e2537eb506e7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5487879
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Sean Harrison <harrisonsean@chromium.org>
Auto-Submit: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/main@{#1292935}