[Signin][Android] Centralize changes to legacy sync account email
This CL replaces assorted calls to setLegacySyncAccountEmail with
calls to maybeUpdateLegacySyncAccountEmail that is now responsible for
keeping the shared pref consistent with the primary account email
used on the native side.

It also promotes the test that verifies the sign-out behavior when
the account is removed to an integration test.

Bug: 337003667
Change-Id: Ic72cd5f456bb664e26494bd28cb7629581d8d936
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5489402
Reviewed-by: Tanmoy Mollik <triploblastic@google.com>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292936}