[FspCache] Evict items if number of items exceeds maximum
The cache has a maximum number of items allowed. When the limit is
reached, the least-recently used should be evicted. This introduces a
new bit on the CacheFileContext to mark files for removal.

A two-stage process is introduced: one to mark items for removal then
another to actually evict these items. The latter is used as the
underlying operations (disk IO and database access) could potentially
cause user visible latency so this will only be performed when
requested.

This also renames SetMaxCacheSize to SetMaxCacheItems to be more
representative of the underlying behavior. Eventually a new method
called SetMaxCacheSizeInBytes will be introduced to indicate the total
allowable size in bytes.

Bug: b/328679426
Test: unit_tests --gtest_filter=FileSystemProviderContentCacheImplTest*
Change-Id: I384fce3181ec148ba6cf18d5ab75b99b152ca59f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5480210
Commit-Queue: Ben Reich <benreich@chromium.org>
Reviewed-by: Cassy Chun-Crogan <cassycc@google.com>
Cr-Commit-Position: refs/heads/main@{#1292944}