Add support to the streaming parser for several utilities.
This is a preparatory change, so that a lot of the Consume*()
functions can work with the streaming parser (largely through
templates). This is dead code now, but it significantly reduces
the delta that we have to keep for the larger patch set.

Bug: 40279985
Change-Id: Ic8f6f9872b65d5a7900ebec142ce33111e8d16df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5488098
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292959}