Rename CSSProperty::ParseSingleValue() to ParseSingleValueFromRange().
This is a preparatory patch before we start making CSS property parsing
start using the streaming parser; for the time being, we will support
both parsing from a stream and from a range, with the former being
the preferred option. Thus, we move names now, in a purely mechanical
change to make the later patches in the series easier to review.

No functional change.

Bug: 40279985
Change-Id: I8ef92a6880265f6bd17d46a79ae8b886550c36b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5457895
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1292961}