Untie popup reopening callback from AutofillContextMenuManager lifespan.
This is a no functional change. Currently, `AutofillContextMenuManager`
is alive after an address is added, and using its weak reference to
get required arguments for autofill popup reopening works. But
there is no documented guarantee for that behaviour. It is more
reliable to not depend on it. This CL does this by removing
the weak pointer and passing the frame and field ids only.

Bug: 325440757
Change-Id: Ief2640247593c9fa4ea46de7768ea23bb6edf760
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5487678
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1293037}