[ios] Update legacy comment in ChromeBrowserStateManager
With old implementation ChromeBrowserState* GetBrowserState(...) was
loading the browserState at path if it was not loaded yet.

With new implementation nullptr is returned is the browserState is not
in the list of loaded browserStates.

The CL updates the obsolete comment.

Bug: none
Change-Id: Ifaee71ea7a5dc1911d256a40001bd7e4f6d78e1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493788
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Auto-Submit: Federica Germinario <fedegermi@google.com>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Cr-Commit-Position: refs/heads/main@{#1293053}