[ios] Improve metrics collection of: IOS.MultiWindow.OpenInNewWindow
This CL improves metrics collection of IOS.MultiWindow.OpenInNewWindow

Metrics are recorded multiple times when a new window is opened,
leading to a wrong collection of WindowActivityUnknownOrigin.
By recording metrics only when level is SceneActivationLevelBackground,
the recording happens only once for each new window creation, and it
will happen with the correct value.

This should reduce significantly number of WindowActivityUnknownOrigin
collected.

This CL is also updating the expiration of the metric.

Fixed: 336510317
Change-Id: Ib89b5064a81a7d0267cd4864d176edf16028c2c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5478067
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Federica Germinario <fedegermi@google.com>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293088}