[CPE] Improve Favicon Fetching Logic
Introduces a refactored AddCredentials method. This method builds a
dictionary of existing favicons and their creation date, and uses it to
determine whether to fetch favicons or not. The initial time threshold
is set to 14 days.

Previously, favicons would be fetched all the time.

Bug: 337268138

Change-Id: I205fbfca68376ca7b0757d7096f1aa9395843891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5462284
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Sebastien Seguin-Gagnon <sebsg@chromium.org>
Reviewed-by: Alexis Hétu <sugoi@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Reviewed-by: Olivier Li <olivierli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293103}