[Shortcuts] Append profile name to initial title in create shortcut UX
This CL appends the user name of the profile to the text field in
the Create Shortcut dialog so that users are aware of which profile
is being used to create the shortcut, if there is more than one profile
in the device at the time of creating the shortcut. The profile name
will be added to the title of the shortcut during shortcut creation, unless the user modifies it.

Screenshot: https://screenshot.googleplex.com/BGvcWofHZkeKPaB

Bug: 336768117, 333024272
Change-Id: I9eb96c1f97474049d4e57f8d81a25f45aa2b569c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5484546
Commit-Queue: Dibyajyoti Pal <dibyapal@chromium.org>
Reviewed-by: Daniel Murphy <dmurph@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293105}