[iOS] Replace raw_ptr with base::WeakPtr in TabGroupItem
This CL replaces raw_ptr with base::WeakPtr in TabGroupItem to refer to
the underlying TabGroup instance. This was necessary to differentiate
between two TabGroupItems referring to two different groups which have
been removed, since there is no such thing as a TabGroupID that can be
used instead. This should not be necessary anymore, since TabGroupItems
now use a `const void* tabGroupIdentifier` to identify themselves, which
does not change when the TabGroup object is removed from memory.

Bug: 336514207, 336513098, 336514873, 336511171
Change-Id: I98508df936b40df4069e278d6cad5f4c4a0c7380
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5480329
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Louis Romero <lpromero@google.com>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1293130}