iwa: Mark isolated-app: as being handled by Chrome
This adds isolated-app: to Chrome's list of handled protocols, which
makes ChildProcessSecurityPolicyImpl::CanRequestURL not always return
true for isolated-app: URLs.

In order for this to not break Client.openWindow in IWA service workers,
this also grants service worker processes access to the worker script's
origin when the process is created.

IsolatedWebAppBrowserServiceWorkerPushTest.
ServiceWorkerPartitionedWhenWakingUpDueToPushNotification tests this
in a browser test.

Bug: b/335511141
Change-Id: Iff834f44f9c520b6d778c679cbed99dea31fe665
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5466063
Commit-Queue: Robbie McElrath <rmcelrath@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293165}