[iOS] Create Async and Sync Url Checker When Async Is Enabled
In this CL, we create an async and sync url checker when the async check feature is enabled. As the feature is disabled by default on iOS, we can add this logic without having a full implementation for doing an async check. The reason we don't include the full implementation for an async check yet is because the logic is rather bulky and it would be better to introduce the changes in smaller CLs.

Bug: 41491791
Change-Id: I4ca4153416ceaf27f7ef9b354c9eca552a7a2eaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5406052
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293181}