view-transition: Cleanup use of tokens to identify ViewTransition.
Each ViewTransition is identified using an UnguessableToken to ensure
it's unique across all processes, since these transitions can be
cross-process for cross-document navigations. Currently this token is
referred to as both viz::TransitionId in contexts which apply to all
transitions; and as viz::NavigationId in contexts specific to
cross-document navigations.

The above pattern is confusing. Switch to using a strongly typed
blink::ViewTransitionToken type throughout the code-base.

Also, this token was currently only being sent through the compositing
stack in transition directives for cross-document navigations. Since the
token will be needed for concurrent transitions within the same
compositor, add it to directives for all transitions. Also add a bit to
indicate when a transition could be cross frame sink (a.k.a cross-document
navigations) for Viz side logic to manage lifetime of cached resources
from the old Document.

Bug: 330917526
Change-Id: Ie81493e3e4ba684766d632c7d33db5759e6f2f65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5491464
Commit-Queue: Khushal Sagar <khushalsagar@chromium.org>
Reviewed-by: Jeremy Roman <jbroman@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Auto-Submit: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293192}