[User Education] Revise IPH validation test.
As a follow-up to the failures in the attached bug (which has now been
fixed), this CL revises the IPH consistency checks:

 - Lack of configuration is no longer an error, as auto-configuration is
   now supported in both V1 and V2.
   * The test still notes that the feature is auto-conf only.
 - Incorrect error feedback has been updated.
 - A more detailed error message is presented which describes what the
   test is for and how to find the offending code causing it to fail.

Before this can be submitted, it should be run on branded Chrome
builders in addition to chromium CQ bots, as some IPH may only be
enabled on branded builds.

The latter two attached bugs are obsoleted by this change.

Bug: 336983096, 40910892, 40910899
Change-Id: Ib295c629f3ae133fbb40d08d8fd8ba0b15f54525
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5492256
Commit-Queue: Dana Fried <dfried@chromium.org>
Reviewed-by: Mickey Burks <mickeyburks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293194}