[Tab Group] Emit token when removing a tab group
When removing a tab group ID whose group ID won't be re-used as the
destination tab group ID emit it as part of `didRemoveTabGroup` and use
that to update sync.

If the ID will be re-used in the newly merged group then the event is
emitted but the tab group ID is null. For all other usages the integer
root ID continues to be emitted.

Fixed: 337135045
Change-Id: Ie01704723581ae6e5228a6b6ec446cc5cfbd5cc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493326
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293206}