[QuickStart]: wait for adapter to be ready after reboot
This CL removes a check inside the bootstrap controller when status
isn't supported. We can handle this gracefully inside the connection
broker. This CL also observes the bluetooth adapter so that advertising
waits for the adapter to be present and powered in the case of a resume
after update.

Test: manually tested with pixel8 and atlas
Change-Id: I756782ae32eb93785a614c2eb6d061b3dba22507
Bug: b/335875866
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493384
Commit-Queue: Brittany Hartmire <bhartmire@google.com>
Reviewed-by: Michael Hansen <hansenmichael@google.com>
Cr-Commit-Position: refs/heads/main@{#1293301}