[TSLO] Check isInDesktopWindow before create new AppHeaderState
crrev.com/c/5486636 Introduced isInDesktopWindow boolean, and AppHeaderState will use this boolean to check if the state is valid. This change exposed assertion error when switching from DW mode into fullscreen, as the AppHeaderState holds a non-empty widestOcculdedRect as bounding rects becomes empty, while the captionBar insets still exists. So we could generate an invalid AppHeaderState.

This CL attempt fix it by checking if the app is in DW mode before creating the AppHeaderState. This ensure the client that receive AppHeaderState will always get an empty widestOcculdedRect is we are not in desktop windowing mode.

Bug: 325351108
Change-Id: Ib3250b73cfac56f3d419623f7cebaf634ccd7c8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5494386
Reviewed-by: Aishwarya Rajesh <aishwaryarj@google.com>
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293303}