snap-group: Fix the gap between divider and snapped window issue
The gap issue was observed between the divider and snapped window,
based on the window hierarchy, the divider position was wrong. This
was due to the window may still be animating and the `edge_gap`
is not corrected calculated.

Demo: http://b/333618907#comment4, http://b/335323278#comment3
Fixed: b/333618907, b/335323278
Test: Added unit tests (both fail w/o changes) + manually
Change-Id: I21124ea301dba4a572e82cb1856a31f83e607f6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5493379
Commit-Queue: Michele Fan <michelefan@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293304}