Be explicit about format of date for history searches
Previously, any string that worked to construct a JS Date object
was okay. This CL makes it so that it only accepts YYYY-MM-DD format.

Also updates `toISOString()` call to manually construct the date
string and updates creating a new date object to include the time
in the string so that both are in the user's timezone not UTC.

Bug: 333936671
Change-Id: I4836e2aeb39261601a4b3a9a32c19fa16634db19
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5492506
Commit-Queue: John Lee <johntlee@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293314}