Reland "wm: Miscellaneous cleanup (1/N)"
This is a reland of commit c737301a3de97c2c496ec00fcacab02208f86b0d

The original patchset is patchset1. I just removed the line that was
causing problems. There seems to be an issue with the default move
assignment. Now that app_restore::WindowInfo has copy and move
assignment operators, it doesn't need to be a pointer anymore. I'll
leave that for a separate change since there are other changes in this
patch and its non trivial.

Original change's description:
> wm: Miscellaneous cleanup (1/N)
>
> - Use EXPECT_THAT for WindowRestoreControllerTest.
> - Use references instead of pointers for output parameters.
> - Remove a no longer valid TODO.
>
> Test: none
> Bug: b/337031769
> Change-Id: I9941e6e65e69b242e9b992ced580494f98b8c5a5
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5484381
> Commit-Queue: Sammie Quon <sammiequon@chromium.org>
> Reviewed-by: Sophie Wen <sophiewen@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1292548}

Test: ran locally with MSAN, used MSAN trybot
Bug: b/337031769
Change-Id: I3db1abbab07cdbcd5d56955d3206c905377729be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5492290
Reviewed-by: Sophie Wen <sophiewen@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293320}