Create a new folder for browser_instance code.
This CL moves all browser_instance code to a new directory in
//chrome/browser/apps. The intention is to allow App Service to depend
on this code unidirectionally, with the only reverse exception being the
test file. This is a step towards simplifying the App Service's BUILD.gn
and dependency tree, which should allow for easier future cleanup of
circular and violated dependencies. There are no functional changes.

As part of the move, the one instance-related method in
web_contents_app_id_utils.cc is moved to a new file in the
browser_instance directory, with dependents updated accordingly. This
separates the instance and app ID util files and splits their
dependencies up.

Bug: b/336172688
Change-Id: I1d367816ec9841cab35477f1dc4320e125c11d0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5473724
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Commit-Queue: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1293512}